Addressed eth2 circular imports (unknown reason)
